Add new `skipDiff` arg to `compareScreenshots`
* Useful when using Cypress retry for snapshot testing commands. By
default there will be a diff on each retry attempt so if the test pass
eventually, there still will be trashy diffs from previous attempts.
With this argument and the knowledge of the try attempt number a bool
condition could be crafted in the user code to allow diff creation only
for the last retry attempt.

* Refactor in `compareScreenshots` to avoid code duplication around diff
creation code

* Allow passthrough of `compareScreenshots` from `compareScreenshots`
command

* Introduce a new Cypress ENV NEVER_GENERATE_DIFF support in case if
users want to permanently disable diffs creation for any reason